requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
gradio>=4.0.0
faiss-cpu>=1.7.0
datasets>=2.0.0
